backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.39
requests==2.32.3
PyJWT==2.10.1
python-dotenv==1.0.1
